requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.6
dj-database-url==0.5.0
dj-stcors-headers==3.4.0
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.11.2
djangorestframework-simplejwt==4.6.0atic==0.0.6
Django==3.1.14
django-
djoser==2.0.3
filelock==3.8.0
gensim==4.3.2
gunicorn==20.1.0
install==1.3.5
Janome==0.4.2
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.1
Pillow==9.0.1
pipenv==2022.9.21
pippip==0.0.10
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-decouple==3.3
pytz==2020.5
PyYAML==6.0.1
render==1.0.0
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
sqlparse==0.4.2
static3==0.7.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
whitenoise==6.3.0
